third_party/blink/web_tests/platform/generic/fast/inline/reattach-inlines-in-anonymous-blocks-with-out-of-flow-siblings-expected.txt

The numbers below should be on the same line.

On success, you will see a series of "
PASS
" messages, followed by "
TEST COMPLETE
".

PASS divs[2].offsetTop is divs[0].offsetTop
PASS successfullyParsed is true

TEST COMPLETE


1 
2 
3
